examples/2-structured-output-libs/pyproject.toml
[project]
name = "1-strctured-output"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "baml-py>=0.206.1",
    "instructor>=1.11.2",
    "jsonschema>=4.25.1",
    "openai>=1.105.0",
    "pydantic-ai>=1.0.0",
    "python-dotenv>=1.1.1",
]
